scripts/sermon-clipper/requirements.txt
Pillow>=10.0.0
webvtt-py>=0.5.0
pysrt>=1.1.0
